requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pymupdf>=1.23.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-chroma>=0.1.0
langchain-google-genai>=0.0.5
langchain-huggingface>=0.0.5
sentence-transformers>=2.2.2
chromadb>=0.4.22
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
requests>=2.31.0
streamlit>=1.30.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
